split up equations plugin
