+++ src/serlo-editor/plugins/equations/editor/editor-renderer.tsx
